Height map uses integer hashing.

+++ SnazzCraft/includes/snazzcraft-engine/height-map/height-map.hpp
@@ -10,13 +10,13 @@ namespace SnazzCraft
     struct HeightMap
     {
     public:
-        std::unordered_map<uint32_t, int32_t> HeightValues = std::unordered_map<uint32_t, int>();
+        std::unordered_map<uint64_t, int32_t> HeightValues = std::unordered_map<uint64_t, int>();
         SnazzCraft::PerlinNoiseModule* NoiseModule = nullptr;
         
         HeightMap(uint32_t Size, int32_t HeightConstraintLow, int32_t HeightConstraintHigh, int32_t Seed, double Frequency, double Persistence, double Lacunarity, int OctaveCount);
         ~HeightMap();
 
-        void GenerateValue(uint32_t X, uint32_t Y);
+        void GenerateValue(int32_t X, int32_t Y);
 
     private:
         uint32_t Size;
